// client/src/app/layouts/full/sidebar/nav-item/nav-item.ts
export interface NavItem {
    displayName?: string;
    iconName?: string;
    navCap?: string;
    route?: string;
    children?: NavItem[];
    chip?: boolean;
    chipContent?: string;
    chipClass?: string;
    external?: boolean;
    bgcolor?:string;
     disabled?: boolean; 
     action?: 'logout' | 'otherAction'; 
}  